SWITCHER-81 Cambios en algunas partes, hay un nuevo archivo donde se puede utilizar el fetch como funcion
