+++ src/components/JoinButton.jsx
@@ -2,6 +2,7 @@ import React from 'react'
 import { Button, Modal} from 'react-bootstrap';
 import { GameContext, GameProvider } from '../contexts/GameContext.jsx';
 import { useState, useContext } from 'react';
+import {usePlayersInfo} from '../utils/fetches.jsx';
 
 function JoinButton ({selectedMatch}) {
     const { idGame, namePlayer, idPlayer, players, fase, setIdGame, setIdPlayer, setNamePlayer, setPlayers, setFase} = useContext(GameContext);
@@ -46,6 +47,7 @@ function JoinButton ({selectedMatch}) {
           setIdPlayer(newIdPlayer);
           setNamePlayer(nickname);
           setFase('lobby');
+          usePlayersInfo();
           console.log(`Joined: ${selectedMatch.id}`)
         } catch (error) {
           console.error('Error:', error);
